Survey model doesn't use knockout any more.
